Implemented all possible messages to be sent from client
