+++ core/src/edu/rice/rbox/Game/Common/Utils/GameClientGrpc.java
@@ -1,5 +1,6 @@
 package edu.rice.rbox.Game.Common.Utils;
 
+import com.google.protobuf.Empty;
 import edu.rice.rbox.Common.GameObjectUUID;
 import edu.rice.rbox.Game.Common.SyncState.PlayerState;
 import io.grpc.ManagedChannel;
@@ -46,6 +47,18 @@ public class GameClientGrpc {
         serverStub.initPlayer(initMessage.getInitialPlayerStateMessage());
 
     return response;
+  }
+
+  public GameNetworkProto.SuperPeerInfo getSuperPeer() {
+    GameNetworkProto.SuperPeerInfo response =
+        serverStub.getAssignedSuperPeer(Empty.newBuilder().build());
+
+    return response;
+  }
+
+  public void remove(GameObjectUUID playerID) {
+    GameNetworkProto.PlayerID.Builder idMessage = GameNetworkProto.PlayerID.newBuilder();
+    serverStub.removeMe(idMessage.setPlayerID(playerID.toString()).build());
   }
 
   public static void main(String args[]) {
