+++ core/src/edu/rice/rbox/Game/Common/Utils/GameGrpc.java
@@ -1,14 +1,7 @@
 package edu.rice.rbox.Game.Common.Utils;
 
-import edu.rice.rbox.Networking.NetworkImpl;
-import io.grpc.Server;
-import io.grpc.ServerBuilder;
+import com.google.protobuf.Empty;
 import io.grpc.stub.StreamObserver;
-import java.io.IOException;
-import java.net.DatagramSocket;
-import java.net.InetAddress;
-import java.net.SocketException;
-import java.net.UnknownHostException;
 import network.GameNetworkProto;
 import network.GameServiceGrpc.GameServiceImplBase;
 
@@ -25,7 +18,7 @@ public class GameGrpc extends GameServiceImplBase {
   }
 
   @Override
-  public void getAssignedSuperPeer(GameNetworkProto.Empty request,
+  public void getAssignedSuperPeer(Empty request,
                                    StreamObserver<GameNetworkProto.SuperPeerInfo> responseObserver) {
   }
 
@@ -37,37 +30,7 @@ public class GameGrpc extends GameServiceImplBase {
   }
 
   @Override
-  public void removeMe(GameNetworkProto.PlayerID request, StreamObserver<GameNetworkProto.Empty> responseObserver) {
-  }
-
-  public static void main(String args[]) throws IOException, InterruptedException {
-    String ip = "";
-    try(final DatagramSocket socket = new DatagramSocket()){
-      socket.connect(InetAddress.getByName("8.8.8.8"), 10002);
-      ip = socket.getLocalAddress().getHostAddress();
-    } catch (SocketException | UnknownHostException e) {
-      System.err.println("Server failed to start.");
-      e.printStackTrace();
-    }
-
-    System.out.println("Server Running on address: " + ip);
-
-
-    // Create a new server to listen on port 8080
-
-    Server server = ServerBuilder.forPort(8080)
-                        .addService(new GameGrpc())
-                        .build();
-
-
-    // Start the server
-    server.start();
-
-    // Server threads are running in the background.
-    System.out.println("Server started");
-
-    // Don't exit the main thread. Wait until server is terminated.
-    server.awaitTermination();
+  public void removeMe(GameNetworkProto.PlayerID request, StreamObserver<Empty> responseObserver) {
   }
 
 }
